difine and deneis parmeeter with current route name intigration

+++ app/Http/Controllers/Admin/UserController.php
@@ -7,11 +7,14 @@ use App\Role;
 use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Route;
 
 class UserController extends Controller
 {
+    private $currentPath;
     public function __construct(){
         $this->middleware('auth');
+        $this->currentPath= Route::getFacadeRoot()->current()->uri();
     }
     /**
      * Display a listing of the resource.
@@ -43,7 +46,7 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        if(Gate::denies('edit-users')){
+        if(Gate::denies($this->currentPath)){
             return redirect()->route('admin.user.index');
         }
         $roles = Role::all();
@@ -71,6 +74,10 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+        if(Gate::denies($this->currentPath)){
+            return redirect()->route('admin.user.index');
+        }
+
         $user->roles()->detach();
         $user->delete();
         return redirect()->route('admin.user.index');
